+++ src/views/Dashboard/CreateDashboard.js
@@ -2,13 +2,16 @@ import React from "react";
 import PropTypes from "prop-types";
 import _ from "lodash";
 import { Responsive, WidthProvider } from "react-grid-layout";
-import { Paper } from '@material-ui/core';
+import { Paper, Dialog, AppBar, IconButton, Toolbar, Hidden, withStyles } from '@material-ui/core';
 import { InfoCard } from 'components';
-
+import { T, ItemG } from 'components';
+import cx from 'classnames'
+import { Close } from 'variables/icons';
+import dashboardStyle from 'assets/jss/material-dashboard-react/dashboardStyle';
 const ResponsiveReactGridLayout = WidthProvider(Responsive);
 
 //ignore
-export default class CreateDashboard extends React.Component {
+class CreateDashboard extends React.Component {
 	constructor(props) {
 		super(props);
 		this.state = {
@@ -84,40 +87,79 @@ export default class CreateDashboard extends React.Component {
 		})
 	}
 	render() {
+		const { openAddDash, handleCloseDT, classes, t } = this.props
+		const appBarClasses = cx({
+			[' ' + classes['primary']]: 'primary'
+		});
+	
 		return (
-			<div>
+			<Dialog
+				fullScreen
+				open={openAddDash}
+				onClose={handleCloseDT}
+				TransitionComponent={this.transition}>
+				<AppBar className={classes.appBar + ' ' + appBarClasses}>
+					<Toolbar>
+						<Hidden mdDown>
+							<ItemG container alignItems={'center'}>
+								<ItemG xs={2} container alignItems={'center'}>
+									<IconButton color='inherit' onClick={handleCloseDT} aria-label='Close'>
+										<Close />
+									</IconButton>
+									<T variant='h6' color='inherit' className={classes.flex}>
+										{t('dashboard.createDashboard')}
+									</T>
+								</ItemG>
+							</ItemG>
+						</Hidden>
+						<Hidden lgUp>
+							<ItemG container alignItems={'center'}>
+								<ItemG xs={4} container alignItems={'center'}>
+									<IconButton color={'inherit'} onClick={handleCloseDT} aria-label='Close'>
+										<Close />
+									</IconButton>
+									<T variant='h6' color='inherit' className={classes.flex}>
+										{t('dashboard.createDashboard')}
+									</T>
+								</ItemG>
+							</ItemG>
+						</Hidden>
+					</Toolbar>
+				</AppBar>
+				<div>
 
-				<div>
+					<div>
 					Current Breakpoint: {this.state.currentBreakpoint} ({
-						this.props.cols[this.state.currentBreakpoint]
-					}{" "}
+							this.props.cols[this.state.currentBreakpoint]
+						}{" "}
 					columns)
-				</div>
-				<div>
+					</div>
+					<div>
 					Compaction type:{" "}
-					{_.capitalize(this.state.compactType) || "No Compaction"}
-				</div>
-				<button onClick={this.handleAddNew}>Add new Column</button>
-				<button onClick={this.onNewLayout}>Generate New Layout</button>
-				<button onClick={this.onCompactTypeChange}>
+						{_.capitalize(this.state.compactType) || "No Compaction"}
+					</div>
+					<button onClick={this.handleAddNew}>Add new Column</button>
+					<button onClick={this.onNewLayout}>Generate New Layout</button>
+					<button onClick={this.onCompactTypeChange}>
 					Change Compaction Type
-				</button>
-				<ResponsiveReactGridLayout
-					{...this.props}
-					layouts={this.state.layouts}
-					onBreakpointChange={this.onBreakpointChange}
-					onLayoutChange={this.onLayoutChange}
-					// WidthProvider option
-					measureBeforeMount={false}
-					// I like to have it animate on mount. If you don't, delete `useCSSTransforms` (it's default `true`)
-					// and set `measureBeforeMount={true}`.
-					useCSSTransforms={this.state.mounted}
-					compactType={this.state.compactType}
-					preventCollision={!this.state.compactType}
-				>
-					{this.generateDOM()}
-				</ResponsiveReactGridLayout>
-			</div>
+					</button>
+					<ResponsiveReactGridLayout
+						{...this.props}
+						layouts={this.state.layouts}
+						onBreakpointChange={this.onBreakpointChange}
+						onLayoutChange={this.onLayoutChange}
+						// WidthProvider option
+						measureBeforeMount={false}
+						// I like to have it animate on mount. If you don't, delete `useCSSTransforms` (it's default `true`)
+						// and set `measureBeforeMount={true}`.
+						useCSSTransforms={this.state.mounted}
+						compactType={this.state.compactType}
+						preventCollision={!this.state.compactType}
+					>
+						{this.generateDOM()}
+					</ResponsiveReactGridLayout>
+				</div>
+			</Dialog>
 		);
 	}
 }
@@ -134,3 +176,4 @@ CreateDashboard.defaultProps = {
 	// initialLayout: generateLayout()
 };
 
+export default withStyles(dashboardStyle)(CreateDashboard)
